feat: add date formatting utility and implement poke functionality
- Create reusable formatDate utility in lib/date.ts
- Replace hardcoded date with dynamic lastOnlineDate formatting
- Implement poke notification API call in member page
- Refactor GroupItem to use shared date utility
