+++ components/group/GroupItem.tsx
@@ -1,6 +1,7 @@
 import { cn } from "@/lib/utils";
 import { BellIcon } from "@/components/icons/BellIcon";
 import { useNewGroupMessages } from "@/api/hooks";
+import formatDate from "@/lib/date";
 
 interface GroupItemProps {
   isJoined: boolean;
@@ -21,13 +22,6 @@ export const GroupItem = ({
   onClick,
   className,
 }: GroupItemProps) => {
-  const formatDate = (dateString: string) => {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    return `${year}.${month}.${day}`;
-  };
   const { data: { hasNewMessages } = { hasNewMessages: false } } =
     useNewGroupMessages(groupId);
   console.log("msg", hasNewMessages);
